🏗️ Move mod apply into JsonSourceCopier
This moves a few layers but not doModProp yet to keep the diffs clean.
